requirements.txt
attrs==23.1.0
Brotli==1.0.9
click==8.1.6
colorama==0.4.6
inflate64==0.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
mimesis==10.1.0
multivolumefile==0.2.3
numpy==1.25.1
psutil==5.9.5
py7zr==0.20.5
pybcj==1.0.1
pycryptodomex==3.18.0
Pygments==2.15.1
pyppmd==1.0.0
python-dateutil==2.8.2
pytz==2023.3
pyzstd==0.15.9
rich==13.4.2
shellingham==1.5.0.post1
six==1.16.0
texttable==1.6.7
typer==0.9.0
types-python-dateutil==2.8.19.13
typing_extensions==4.7.1
tzdata==2023.3
XlsxWriter==3.1.2
